Lighthouse pushing to network

+++ web3-lighthouse-keeper/src/main/java/web3/lighthouse/keeper/model/CoinData.java
@@ -1,5 +1,7 @@
 package web3.lighthouse.keeper.model;
 
+import java.math.BigInteger;
+
 public interface CoinData {
-  int getPrice();
+  BigInteger getPrice();
 }
